requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via geoip2
appdirs==1.4.4
    # via urlextract
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
certifi==2020.12.5
    # via requests
chardet==3.0.4
    # via
    #   -r requirements.in
    #   aiohttp
    #   requests
colorama==0.4.4
    # via -r requirements.in
confluent-kafka==1.5.0
    # via -r requirements.in
geoip2==4.1.0
    # via -r requirements.in
idna-ssl==1.1.0
    # via aiohttp
idna==2.10
    # via
    #   -r requirements.in
    #   idna-ssl
    #   requests
    #   tldextract
    #   urlextract
    #   yarl
jsonref==0.2
    # via -r requirements.in
luqum==0.11.0
    # via -r requirements.in
maxminddb==2.0.3
    # via geoip2
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.19.5
    # via -r requirements.in
ply==3.11
    # via luqum
pycryptodome==3.10.1
    # via -r requirements.in
pygrok==1.0.0
    # via -r requirements.in
python-dateutil==2.8.1
    # via -r requirements.in
pytz==2021.1
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
regex==2021.3.17
    # via pygrok
requests-file==1.5.1
    # via tldextract
requests==2.25.1
    # via
    #   geoip2
    #   requests-file
    #   tldextract
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.16.13
    # via -r requirements.in
six==1.15.0
    # via
    #   python-dateutil
    #   requests-file
tldextract==2.2.2
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   yarl
ujson==3.1.0
    # via -r requirements.in
uritools==3.0.1
    # via urlextract
urlextract==0.13.0
    # via -r requirements.in
urllib3==1.26.4
    # via
    #   geoip2
    #   requests
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
